refactor: move main function to separate module
